Code optimiz
Remove unused route from app
Header and fact are now component.
Largely decrease the amount of code in index page.
Index page now only load component.
Refactor Navbar so it is ready for future dev.
Replace some image in intro page.
